Temporarily remove non-validated examples
This commit removes examples that have not yet been converted or
validated to work with controller v3. These examples will REMAIN
accessible in the controller-v2 branch of the repository, and next steps
will be to continue to update and validate them one-by-one, then add
them back to the main branch once we have.